serialserver/pyproject.toml
[project]
name = "serialserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pyserial>=3.5",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.1",
]
